defer crowding and new cars calculation until it's needed
