+++ test/content/messages/predictions_test.exs
@@ -1,17 +1,7 @@
 defmodule Content.Message.PredictionsTest do
   use ExUnit.Case, async: true
-  import ExUnit.CaptureLog
-  import Mox
-
-  @sign %{source_config: nil, location_engine: Engine.Locations.Mock}
 
   describe "non_terminal/3" do
-    setup do
-      stub(Engine.Locations.Mock, :for_vehicle, fn _ -> nil end)
-
-      :ok
-    end
-
     test "puts ARR on the sign when train is 0 seconds away, but not boarding" do
       prediction = %Predictions.Prediction{
         seconds_until_arrival: 0,
@@ -22,7 +12,7 @@ defmodule Content.Message.PredictionsTest do
         destination_stop_id: "70261"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont        ARR"
     end
@@ -37,7 +27,7 @@ defmodule Content.Message.PredictionsTest do
         boarding_status: "Boarding"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont        BRD"
     end
@@ -52,7 +42,7 @@ defmodule Content.Message.PredictionsTest do
         destination_stop_id: "70275"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Mattapan       ARR"
     end
@@ -67,7 +57,7 @@ defmodule Content.Message.PredictionsTest do
         destination_stop_id: "70275"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Mattapan     1 min"
     end
@@ -82,7 +72,7 @@ defmodule Content.Message.PredictionsTest do
         destination_stop_id: "70275"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Mattapan     1 min"
     end
@@ -98,7 +88,7 @@ defmodule Content.Message.PredictionsTest do
         destination_stop_id: "70275"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Mattapan   20+ min"
     end
@@ -113,7 +103,7 @@ defmodule Content.Message.PredictionsTest do
         destination_stop_id: "70261"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont      1 min"
     end
@@ -128,7 +118,7 @@ defmodule Content.Message.PredictionsTest do
         destination_stop_id: "70261"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont      2 min"
     end
@@ -143,7 +133,7 @@ defmodule Content.Message.PredictionsTest do
         destination_stop_id: "70261"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont        ARR"
     end
@@ -158,7 +148,7 @@ defmodule Content.Message.PredictionsTest do
         boarding_status: "Boarding"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont        BRD"
     end
@@ -173,7 +163,7 @@ defmodule Content.Message.PredictionsTest do
         destination_stop_id: "70261"
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont      2 min"
     end
@@ -188,7 +178,7 @@ defmodule Content.Message.PredictionsTest do
         stops_away: 2
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "RJFK", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "RJFK", "m")
 
       assert Content.Message.to_string(msg) == [
                {"Alewife (A)  5 min", 6},
@@ -206,7 +196,7 @@ defmodule Content.Message.PredictionsTest do
         stops_away: 2
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "RJFK", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "RJFK", "m")
 
       assert Content.Message.to_string(msg) == [
                {"Alewife (B)  5 min", 6},
@@ -224,7 +214,7 @@ defmodule Content.Message.PredictionsTest do
         stops_away: 2
       }
 
-      msg = Content.Message.Predictions.non_terminal(prediction, "RJFK", "m", @sign)
+      msg = Content.Message.Predictions.non_terminal(prediction, "RJFK", "m")
 
       assert Content.Message.to_string(msg) == [
                {"Alewife      6 min", 6},
@@ -234,12 +224,6 @@ defmodule Content.Message.PredictionsTest do
   end
 
   describe "terminal/3" do
-    setup do
-      stub(Engine.Locations.Mock, :for_vehicle, fn _ -> nil end)
-
-      :ok
-    end
-
     test "puts boarding on the sign when train is supposed to be boarding according to rtr" do
       prediction = %Predictions.Prediction{
         seconds_until_departure: 75,
@@ -251,7 +235,7 @@ defmodule Content.Message.PredictionsTest do
         boarding_status: "Stopped at station"
       }
 
-      msg = Content.Message.Predictions.terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont        BRD"
     end
@@ -267,7 +251,7 @@ defmodule Content.Message.PredictionsTest do
         boarding_status: "Stopped at station"
       }
 
-      msg = Content.Message.Predictions.terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont      1 min"
     end
@@ -283,7 +267,7 @@ defmodule Content.Message.PredictionsTest do
         boarding_status: "Stopped at station"
       }
 
-      msg = Content.Message.Predictions.terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont      2 min"
     end
@@ -298,7 +282,7 @@ defmodule Content.Message.PredictionsTest do
         destination_stop_id: "70261"
       }
 
-      msg = Content.Message.Predictions.terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == "Ashmont      1 min"
     end
@@ -313,7 +297,7 @@ defmodule Content.Message.PredictionsTest do
         stops_away: 0
       }
 
-      msg = Content.Message.Predictions.terminal(prediction, "test", "m", @sign)
+      msg = Content.Message.Predictions.terminal(prediction, "test", "m")
 
       assert Content.Message.to_string(msg) == [
                {"Oak Grove    2 min", 6},
